Version 3.0

+++ EDesign/Inc/user.h
@@ -26,7 +26,7 @@ void User(void);
 void resetAll(void);
 void writeToPins(uint8_t segments[], uint8_t pins[], int segmentsL);
 
-uint8_t Int2String(char* outputString, int16_t value, uint8_t maxL);
+uint8_t Int2String(char* outputString, int32_t value, uint8_t maxL);
 uint8_t String2Int(char* inputString, int16_t* outputInt);
